fix(app): address CI and review feedback
