+++ packages/app/src/app/transaction-fee.tsx
@@ -1,41 +1,15 @@
-import { Stack, useRouter } from 'expo-router';
-import { useEffect, useRef } from 'react';
+import { Stack } from 'expo-router';
 import { View } from 'react-native';
 
-import { isDefined } from '@rnw-community/shared';
-
-import { useFormsheetListStyles } from '../@generic/hook/use-formsheet-list-styles/use-formsheet-list-styles.hook';
+import { useModalRouteState } from '../@generic/hook/use-modal-route-state/use-modal-route-state.hook';
 import { TransactionFeeModalContent } from '../transaction/components/transaction-fee-modal-content/transaction-fee-modal-content';
 import { useTransactionFeeModal } from '../transaction/context/transaction-fee-modal.context';
 
 export default function TransactionFeeModal() {
-    const router = useRouter();
     const [, resolveTransactionFee, currentParams] = useTransactionFeeModal();
-    const { backgroundColor } = useFormsheetListStyles();
-    const hadParamsRef = useRef(isDefined(currentParams));
+    const { backgroundColor, screenOptions } = useModalRouteState(currentParams, resolveTransactionFee, null);
 
-    const screenOptions = { contentStyle: { backgroundColor } };
     const containerStyle = { backgroundColor };
-
-    useEffect(
-        () => () => {
-            resolveTransactionFee(null, { skipBack: true });
-        },
-        // eslint-disable-next-line react-hooks/exhaustive-deps -- Resolve on unmount only
-        []
-    );
-
-    useEffect(() => {
-        if (isDefined(currentParams)) {
-            hadParamsRef.current = true;
-
-            return;
-        }
-
-        if (!hadParamsRef.current) {
-            router.back();
-        }
-    }, [currentParams, router]);
 
     if (!currentParams) {
         return null;
