+++ packages/app/src/transaction/components/refunded-pill/refunded-pill.tsx
@@ -2,19 +2,16 @@ import { TransactionConsolidationTypeEnum, TransactionEntryTypeEnum, UserIconNam
 import { getLogger } from '@budgie/logger';
 import { t } from '@lingui/core/macro';
 import { useEffect, useState } from 'react';
-import { Text, View } from 'react-native';
-import Animated, { FadeIn } from 'react-native-reanimated';
 
 import { getErrorMessage, isDefined, isNotEmptyString } from '@rnw-community/shared';
 
-import { HapticPressable } from '../../../@generic/component/haptic-pressable/haptic-pressable';
-import { Icon } from '../../../@generic/component/icon/icon';
 import { transactionRepository } from '../../../@generic/drizzle/db/db';
 import { convertFromMicroUnits } from '../../../@generic/utils/convert-from-micro-units.util';
 import { useFormatDigits } from '../../../i18n/hook/use-format-digits.hook';
 import { useSettingsContext } from '../../../settings/context/settings.context';
 import { RefundedSummaryKindEnum } from '../../enum/refunded-summary-kind.enum';
 import { computeRefundedSummary } from '../../utils/compute-refunded-summary.util';
+import { TransactionMetaPill } from '../transaction-meta-pill/transaction-meta-pill';
 
 import type { TransactionWithRelationsEntityInterface } from '@budgie/contracts';
 
@@ -24,11 +21,6 @@ interface Props {
     readonly testID?: string;
 }
 
-const PILL_ICON_SIZE = 12;
-const ENTER_DURATION_MS = 200;
-const baseClassName = 'flex-row items-center gap-x-xs rounded-full px-sm py-[2px] border border-secondary-corner bg-secondary-background';
-const labelClassName = 'text-xs text-secondary-foreground';
-const continuousBorder = { borderCurve: 'continuous' as const };
 const logger = getLogger('RefundedPill');
 
 export const RefundedPill = ({ transaction, onPress, testID }: Props) => {
@@ -83,32 +75,5 @@ export const RefundedPill = ({ transaction, onPress, testID }: Props) => {
             : null;
     const label = isNotEmptyString(formattedRefundedAmount) ? t`Refunded ${formattedRefundedAmount}` : t`Refunded`;
 
-    const body = (
-        <View className={baseClassName} style={continuousBorder}>
-            <Icon icon={UserIconNameEnum.RotateCcw} size={PILL_ICON_SIZE} className={labelClassName} />
-            <Text className={labelClassName} numberOfLines={1} ellipsizeMode="tail">
-                {label}
-            </Text>
-        </View>
-    );
-
-    if (!isDefined(onPress)) {
-        return (
-            <View className="flex-row">
-                <Animated.View entering={FadeIn.duration(ENTER_DURATION_MS)} testID={testID}>
-                    {body}
-                </Animated.View>
-            </View>
-        );
-    }
-
-    return (
-        <View className="flex-row">
-            <Animated.View entering={FadeIn.duration(ENTER_DURATION_MS)}>
-                <HapticPressable accessibilityLabel={label} accessibilityRole="button" accessible onPress={onPress} testID={testID}>
-                    {body}
-                </HapticPressable>
-            </Animated.View>
-        </View>
-    );
+    return <TransactionMetaPill label={label} icon={UserIconNameEnum.RotateCcw} onPress={onPress} testID={testID} />;
 };
